requirements.txt
Flask
msal
requests
python-pptx
python-dotenv